feat: add cjs to esm
